refactor(attendance): split WeekStatisticsCard into separate components
  - Extract navigation controls into AttendanceHeaderControls component
  - Extract statistics display into AttendanceStatisticsCard component
  - Improve view hierarchy with better component separation
  - fix(auth): remove incorrect 'auth/' prefix from API endpoints
